to Siteground FINAL YIPEEEE

+++ deleteLocation.php
@@ -2,21 +2,43 @@
     session_cache_expire(30);
     session_start();
 
-    if ($_SESSION['access_level'] < 2 || $_SERVER['REQUEST_METHOD'] != 'POST') {
+    if ($_SESSION["system_type"] == "MedTracker") {
+        if ($_SESSION['access_level'] < 2 || $_SERVER['REQUEST_METHOD'] != 'POST') {
+            header('Location: index.php');
+            die();
+        }
+        require_once('database/dbServices.php');
+        require_once('include/input-validation.php');
+        $args = sanitize($_POST);
+        $id = $args['id'];
+        if (!$id) {
+            header('Location: index.php');
+            die();
+        }
+        if (delete_location($id)) {
+            header('Location: index.php?deleteLocation');
+            die();
+        }
         header('Location: index.php');
-        die();
+    } else {
+        if ($_SESSION['access_level'] < 2 || $_SERVER['REQUEST_METHOD'] != 'POST') {
+            header('Location: VMS_index.php');
+            die();
+        }
+        require_once('database/dbServices.php');
+        require_once('include/input-validation.php');
+        $args = sanitize($_POST);
+        $id = $args['id'];
+        if (!$id) {
+            header('Location: VMS_index.php');
+            die();
+        }
+        if (delete_location($id)) {
+            header('Location: VMS_index.php?deleteLocation');
+            die();
+        }
+        header('Location: VMS_index.php');
+
+
     }
-    require_once('database/dbServices.php');
-    require_once('include/input-validation.php');
-    $args = sanitize($_POST);
-    $id = $args['id'];
-    if (!$id) {
-        header('Location: index.php');
-        die();
-    }
-    if (delete_location($id)) {
-        header('Location: index.php?deleteLocation');
-        die();
-    }
-    header('Location: index.php');
 ?>
